Port THelpMgr turn-state advisory cluster (0x501270 90.9%, 0x501a20 + 0x5011a0 100%) and rewrite the TGreatPower message builders
- 0x501270 THelpMgr::DispatchTurnStateSpecialAdvisoriesAndReturnCount (1564B, 90.9%):
  full turn-event advisory dispatcher — per-event advisories for queue codes
  0x13b/0x13a/0x131 (nation/minor name formatting, string group 0x2753 templates,
  scanBracketExpressions substitution, localized message dispatch), the map-context and
  town-list message summaries, and the city production reminder. Promoted from an
  anonymous return-0 stub to the real THelpMgr member (all orig call sites load ECX).
- 0x501a20 THelpMgr::ShowPeriodicCapabilityReminderIfNeeded (342B): 100% match.
- 0x5011a0 THelpMgr::HandlePostDispatchTurnStateEventUpdates: 19% -> 100% — early-return
  restructure, removed an invented null check, localization helpers made __inline
  (orig inlines the g_pSimMgr field reads), tick fetched twice as in the orig.
- 0x501be0 ShowPeriodicNationComparisonAdvisoryIfNeeded promoted to THelpMgr member.
- TGreatPower slots 0x53/0x54 retyped void() -> char(CString* outMessageText) (RET 4 +
  AL read + &CString push at every call site) and both bodies rewritten to the real
  logic: append "\n     <name>" lines per qualifying map zone / unlinked town
  (0x4dc660 38%->85%, 0x4dc840 39%->95%).
- TDiplomacyMgr slots 0x44/0x70 param types int -> short (callers push raw 16-bit
  loads; impls movsx the low word); IsNationPairAtWar back to 84.6%,
  GetNationPairDiplomacyRelationCode 62.5% -> 93.3% (row-pointer indexing form).
- TAssetMgr::BuildScenarioPathForModeAndIndex tightened (g_szDecimalFormat + named
  copy local, 85.2%); g_pszEmptyTextPointer_00656f60 recovered as a const pointer
  global to g_szEmptyString.

Verified: build clean, gates pass (datacmp baseline ratcheted: 2 layout-wobble
globals + the new reloc-differing pointer global), 109 tooling tests OK, stats +31
improved; 0x5ba0e0/0x4df5f0 drops are addressing-form/TU wobble in untouched code
(same ops, shifted anchors) accepted per the score-wobble policy.

+++ include/game/TDiplomacyMgr.h
@@ -37,7 +37,7 @@ public:
   virtual void RebuildDiplomacyStandingAndInfluenceMatrices();                  // 14 (0x38)
   virtual void InitializeDiplomacyStandingBaselineRandom();                     // 15 (0x3c)
   virtual void BuildMajorNationDiplomacyStandingRanking();                      // 16 (0x40)
-  virtual char IsNationPairAtWar(int sourceNation, int targetNation);           // 17 (0x44)
+  virtual char IsNationPairAtWar(short sourceNation, short targetNation);       // 17 (0x44)
   virtual char IsNationPairRelationTurnStampOutOfDate(int sourceNation,
                                                       int targetNation);       // 18 (0x48)
   virtual char HasAnyWarRelationForNation(int sourceNation);                   // 19 (0x4c)
@@ -53,8 +53,8 @@ public:
                                                      int targetNation); // 26 (0x68)
   virtual void ShowRelationCodeNoticeForNationPairIfRelevant(int sourceNation,
                                                              int targetNation); // 27 (0x6c)
-  virtual short GetNationPairDiplomacyRelationCode(int sourceNation,
-                                                   int targetNation); // 28 (0x70)
+  virtual short GetNationPairDiplomacyRelationCode(short sourceNation,
+                                                   short targetNation); // 28 (0x70)
   virtual void SetNationPairDiplomacyRelationCode(int sourceNation, int targetNation,
                                                   int relationCode, int updateMode); // 29 (0x74)
   virtual void SetNationPairDiplomacyRelationCodeFinal(int sourceNation, int targetNation,
@@ -77,7 +77,7 @@ public:
   WrapperFor_IsNationSlotEligibleForEventProcessingAt413250(int nationSlot); // 39 (0x9c)
 
   // Slot-name aliases retained for recovered call sites.
-  char HasPolicyWithNationSlot44(int sourceNation, int targetNation) {
+  char HasPolicyWithNationSlot44(short sourceNation, short targetNation) {
     return IsNationPairAtWar(sourceNation, targetNation);
   }
   char HasOutdatedWarRelationSlot48(int sourceNation, int targetNation) {
@@ -90,7 +90,7 @@ public:
   int GetRelationTypeSlot68(int sourceNation, int targetNation) {
     return GetNationPairDiplomacyStandingTierCode(sourceNation, targetNation);
   }
-  short GetRelationTierSlot70(int sourceNation, int targetNation) {
+  short GetRelationTierSlot70(short sourceNation, short targetNation) {
     return GetNationPairDiplomacyRelationCode(sourceNation, targetNation);
   }
   void SetRelationCodeSlot74WithMode(int sourceNation, int targetNation, int relationCode,
